requirements.txt
django==3.2.11
pymysql==1.0.2
django-ratelimit==1.0.1